# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shivanshu Tiwari Portfolio API",
    "version": "1.0.0",
    "description": "Official programmatic REST API for Shivanshu Tiwari — AI agent systems, autonomous evaluation, and backend architectures. Includes async jobs, sandbox testing, and OpenAPI service descriptions.",
    "contact": {
      "name": "Shivanshu Tiwari",
      "url": "https://shivanshutiwari.in/contact",
      "email": "sht4bharat@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/SHT4BHARAT/NoIntroNeeded/blob/main/LICENSE"
    }
  },
  "servers": [
    { "url": "https://shivanshutiwari.in", "description": "Production API server" },
    { "url": "https://shivanshutiwari.in/api/v1/sandbox", "description": "Sandbox test environment" }
  ],
  "x-async-jobs": {
    "dispatch": "/api/v1/jobs",
    "status": "/api/v1/jobs/{jobId}"
  },
  "x-deprecation-policy": {
    "noticePeriodDays": 180,
    "sunsetHeader": "Sunset",
    "deprecationHeader": "Deprecation",
    "policyUrl": "https://shivanshutiwari.in/developers/deprecation"
  },
  "paths": {
    "/api/v1/projects": {
      "get": {
        "operationId": "listProjects",
        "summary": "List portfolio projects with cursor pagination",
        "description": "Returns paginated engineering projects. Responses include RFC 8594 Deprecation and Sunset headers.",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }, "description": "Page size" },
          { "name": "cursor", "in": "query", "schema": { "type": "string" }, "description": "Opaque pagination cursor" },
          { "name": "domain", "in": "query", "schema": { "type": "string", "enum": ["ai-agents", "backend", "voice-ai", "rl"] }, "description": "Filter by technical domain" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of projects",
            "headers": {
              "RateLimit-Limit": { "schema": { "type": "integer" } },
              "RateLimit-Remaining": { "schema": { "type": "integer" } },
              "RateLimit-Reset": { "schema": { "type": "integer" } },
              "Deprecation": { "schema": { "type": "string" }, "description": "RFC 8594 deprecation timestamp" },
              "Sunset": { "schema": { "type": "string" }, "description": "RFC 8594 sunset HTTP-date" },
              "Link": { "schema": { "type": "string" }, "description": "Link to deprecation policy" }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "pagination"],
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Project" } },
                    "pagination": {
                      "type": "object",
                      "required": ["nextCursor", "hasMore"],
                      "properties": {
                        "nextCursor": { "type": "string", "nullable": true },
                        "hasMore": { "type": "boolean" },
                        "limit": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{slug}": {
      "get": {
        "operationId": "getProject",
        "summary": "Get project details by slug",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Project slug" }
        ],
        "responses": {
          "200": {
            "description": "Project details",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } }
          },
          "404": {
            "description": "Not found",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/v1/contact": {
      "post": {
        "operationId": "submitContact",
        "summary": "Submit contact inquiry or internship opportunity",
        "parameters": [
          { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Client-supplied idempotency key" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "message"],
                "properties": {
                  "name": { "type": "string", "example": "Agent Aria" },
                  "email": { "type": "string", "format": "email", "example": "aria@example.com" },
                  "message": { "type": "string", "minLength": 10, "example": "Internship opportunity inquiry" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inquiry received",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "success": { "type": "boolean" }, "message": { "type": "string" } }
                }
              }
            }
          },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "429": { "description": "Rate limited", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/v1/jobs": {
      "post": {
        "operationId": "createJob",
        "summary": "Dispatch asynchronous long-running task",
        "description": "Dispatches an asynchronous processing job. Returns HTTP 202 Accepted with a Location header pointing to the job polling endpoint.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "task": { "type": "string", "example": "benchmark-evaluation" },
                  "payload": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job accepted for asynchronous processing. Poll the URL in the Location header for status.",
            "headers": {
              "Location": {
                "description": "URL to poll for job status and output",
                "schema": { "type": "string", "format": "uri" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Job" }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{jobId}": {
      "get": {
        "operationId": "getJobStatus",
        "summary": "Poll status and results of an asynchronous job",
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Unique job identifier" }
        ],
        "responses": {
          "200": {
            "description": "Current status and output of the job",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Job" }
              }
            }
          }
        }
      }
    },
    "/api/v1/sandbox/ping": {
      "get": {
        "operationId": "sandboxPing",
        "summary": "Sandbox environment health and status check",
        "description": "Verify sandbox test environment connectivity without modifying production data.",
        "responses": {
          "200": {
            "description": "Sandbox status",
            "headers": {
              "X-Sandbox": { "schema": { "type": "string", "enum": ["true"] }, "description": "Indicates sandbox mode" }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["pong", "sandbox"],
                  "properties": {
                    "pong": { "type": "boolean", "example": true },
                    "sandbox": { "type": "boolean", "example": true }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sandbox/contact": {
      "post": {
        "operationId": "sandboxContact",
        "summary": "Simulate contact submission in sandbox mode",
        "description": "Validates schema and idempotency without delivering external notifications.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "message"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "message": { "type": "string", "minLength": 10 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Simulated contact receipt",
            "headers": { "X-Sandbox": { "schema": { "type": "string", "enum": ["true"] } } },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "success": { "type": "boolean" }, "sandbox": { "type": "boolean" } }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/keys": {
      "post": {
        "operationId": "generateTestKey",
        "summary": "Generate an ephemeral sandbox API key for automated testing",
        "responses": {
          "200": {
            "description": "Generated test API key",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["apiKey", "tier", "expiresAt"],
                  "properties": {
                    "apiKey": { "type": "string", "example": "sht_test_sandbox_12345" },
                    "tier": { "type": "string", "example": "sandbox" },
                    "expiresAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/batch": {
      "post": {
        "operationId": "batchOperations",
        "summary": "Execute bulk or batch operations",
        "description": "Bulk endpoint accepting an array of operations or requests to execute in a single round-trip.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["operations"],
                "properties": {
                  "operations": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["op", "id"],
                      "properties": {
                        "op": { "type": "string", "example": "get_project" },
                        "id": { "type": "string", "example": "op_1" },
                        "params": { "type": "object" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch operation execution results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["results"],
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["id", "status"],
                        "properties": {
                          "id": { "type": "string" },
                          "status": { "type": "string", "example": "ok" },
                          "data": { "type": "object" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/batch": {
      "post": {
        "operationId": "batchOperationsAlias",
        "summary": "Execute bulk or batch operations (root alias)",
        "description": "Root alias for bulk operations accepting an array of operations.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "operations": { "type": "array", "items": { "type": "object" } }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch operation execution results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Job": {
        "type": "object",
        "required": ["jobId", "status", "pollUrl"],
        "properties": {
          "jobId": { "type": "string", "example": "job_1725450000" },
          "status": { "type": "string", "enum": ["pending", "running", "completed", "failed"], "example": "pending" },
          "pollUrl": { "type": "string", "format": "uri", "example": "https://shivanshutiwari.in/api/v1/jobs/job_1725450000" },
          "createdAt": { "type": "string", "format": "date-time" },
          "result": { "type": "object", "nullable": true }
        }
      },
      "Project": {
        "type": "object",
        "required": ["slug", "title", "description"],
        "properties": {
          "slug": { "type": "string", "example": "agentic-honey-pot" },
          "title": { "type": "string", "example": "Agentic Honeypot" },
          "description": { "type": "string" },
          "stack": { "type": "array", "items": { "type": "string" } },
          "domain": { "type": "string", "enum": ["ai-agents", "backend", "voice-ai", "rl"] }
        }
      },
      "Error": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": { "type": "string", "example": "not_found" },
          "message": { "type": "string", "example": "Project not found" },
          "hint": { "type": "string", "example": "Check /openapi.json" },
          "requestId": { "type": "string" }
        }
      }
    }
  }
}